+++ src/lib/acl.js
@@ -64,6 +64,7 @@ class Acl {
     this.logger = getLogger('acl');
     this.enabled = config.acl.enabled !== undefined ? config.acl.enabled : false;
     this.cache = new AclCache();
+    this.ALLOWED_PERMISSIONS = new Set(['read', 'write', 'admin']);
   }
 
   /**
@@ -166,24 +167,17 @@ class Acl {
       throw new Error('filePath, permission and dbClient are required');
     }
 
-    let cached = this.cache.getDirPermissions(opts.requestor || 'PUBLIC', filePath, permission);
-    if( cached !== null ) return cached;
-
-    let permissionFilter;
-    if( !opts.requestor ) {
-      // cheat and only allow read access for public if no user is provided
-      if( permission !== 'read' ) return false;
-      permissionFilter = `can_read = true`;
-    } else if( permission === 'read' ) {
-      permissionFilter = `can_read = true`;
-    } else if( permission === 'write' ) {
-      permissionFilter = `can_write = true`;
-    } else if( permission === 'admin' ) {
-      permissionFilter = `is_admin = true`;
-    } else {
+    if( !this.ALLOWED_PERMISSIONS.has(permission) ) {
       throw new AclAccessError('Invalid permission', opts.requestor, filePath, permission);
     }
 
+    let cached = this.cache.getDirPermissions(opts.requestor || 'PUBLIC', filePath, permission);
+    if( cached !== null ) return cached;
+
+    if( !opts.requestor ) {
+      // cheat and only allow read access for public if no user is provided
+      if( permission !== 'read' ) return false;
+    }
     let args = [filePath];
 
     // handle user being null (public access)
@@ -235,15 +229,28 @@ class Acl {
       )`);
     }
 
+    console.log(`
+      WITH ${withQueries.join(', ')}
+      SELECT * FROM ${config.database.schema}.has_permission(
+        (SELECT directory_id FROM dir),
+        '${permission}',
+        ${user ? `(SELECT user_id FROM acluser)` : 'NULL'}
+      )
+    `);
     let resp = await dbClient.query(`
       WITH ${withQueries.join(', ')}
-      SELECT * from ${config.database.schema}.directory_user_permissions_lookup
-      WHERE directory_id = (SELECT directory_id FROM dir)
-      AND ${userSelectQuery}
-      AND ${permissionFilter}
-      `, args);
+      SELECT * FROM ${config.database.schema}.has_permission(
+        (SELECT directory_id FROM dir)::UUID,
+        '${permission}',
+        ${user ? `(SELECT user_id FROM acluser)::UUID` : 'NULL'}
+      )
+    `, args);
 
-    let value = resp.rows.length > 0;
+    let value = false;
+    if( resp.rows.length > 0 ) {
+      value = resp.rows[0].has_permission;
+    }
+
     this.cache.setDirPermissions(opts.requestor || 'PUBLIC', filePath, permission, value);
     return value;
   }
@@ -779,27 +786,6 @@ class Acl {
       }
       await this.setDirectoryAcl({ directoryId: child.directory_id, rootDirectoryAclId, dbClient });
     }
-  }
-
-  /**
-   * @method refreshLookupTable
-   * @description Refresh the directory_user_permissions_lookup materialized view.
-   * This should be run after any changes to ACLs or permissions.  The view is used
-   * for permission checks and is refreshed concurrently to avoid locking but is required
-   * to ensure all query ACL checks are up to date.
-   *
-   * @param {Object} opts
-   * @param {Object} opts.dbClient Required. database client instance
-   * 
-   * @returns {Promise<void>}
-   */
-  refreshLookupTable(opts={}) {
-    let { dbClient } = opts;
-    if( !dbClient ) {
-      throw new Error('dbClient is required');
-    }
-
-    return dbClient.query(`REFRESH MATERIALIZED VIEW CONCURRENTLY ${config.database.schema}.directory_user_permissions_lookup`);
   }
 
 }
