+++ src/lib/transfer.js
@@ -7,7 +7,6 @@ import { pipeline } from 'stream/promises';
 import config from './config.js';
 import { getLogger } from './logger.js';
 import { DuplicateFileError } from './errors.js';
-import Database from './database/index.js';
 import { createContext } from './context.js';
 
 const RDF_MIME_TYPES = new Set([
@@ -317,300 +316,6 @@ class Transfer {
       Buffer.from(JSON.stringify(buckets.rows, null, 2)));
   }
 
-  // ---------------------------------------------------------------------------
-  // IMPORT
-  // ---------------------------------------------------------------------------
-
-  /**
-   * @method prepareImport
-   * @description Import CAS files and optional ACL / auto-partition data from a .tar.gz archive
-   * produced by export(). This method extracts the archive and processes each entry and
-   * provides a callback of batches of files to be imported via optimistic write impl then
-   * 
-   * 
-   * @param {String|import('stream').Readable} src - absolute file path to read from, or a
-   *   Readable stream to consume (e.g. an HTTP request body)
-   * @param {Object} [opts={}]
-   * @param {Boolean} [opts.overwrite=false] - replace existing file records; default throws DuplicateFileError
-   * @param {String} [opts.aclConflict='fail'] - 'fail' | 'skip' | 'merge' on ACL conflicts
-   * @param {String} [opts.autoPartitionConflict='fail'] - 'fail' | 'skip' | 'merge' on rule conflicts
-   * @param {Function} [opts.cb] - optional progress callback `({type, current, total, hash}) => {}`
-   *
-   * @returns {Promise<{hashCount: number, fileCount: number, skippedFiles: number}>}
-   */
-  async import(src, opts={}) {
-    const input = typeof src === 'string' ? fs.createReadStream(src) : src;
-
-
-    const aclData = {};
-    const autoPartitionData = {};
-
-    const batchSize = opts.batchSize || config.sync.defaultBatchSize;
-    const cask = opts.cask;
-    const stats = {
-      filesProcessed: 0,
-      filesInserted : 0,
-      metadataUpdates : 0,
-      noChanges : 0,
-      errors: 0
-    };
-
-    if( !cask ) {
-      throw new Error('CaskFS client instance is required for import');
-    }
-    if( !opts.dbClient ) {
-      throw new Error('Database connection is required for import');
-    }
-
-    // One dedicated DB connection for the entire import operation.
-    const db = opts.dbClient;
-    await db.connect();
-
-    // extract location
-    let extractLocation = opts.extractLocation;
-    if( !extractLocation ) {
-      let parts = path.parse(src);
-      extractLocation = path.join(parts.dir, '._'+parts.name+'_extracted');
-    }
-    if( fs.existsSync(extractLocation) ) {
-      await fs.promises.rm(extractLocation, { recursive: true });
-    }
-    await fs.promises.mkdir(extractLocation, { recursive: true });
-    if( opts.cb ) {
-      opts.cb({ type: 'tmp-file', path: extractLocation });
-    }
-
-    // now extract the archive, processing each entry as it comes.  The .json metadata entries
-    // will come after their corresponding raw files, so the hash → size info will be available
-    // in the rawFileCache when we process the .json entries.
-    let files;
-    if( fstatSync(input.fd).isFile() ) {
-      files = await this._extractArchiveToLocation(input, extractLocation);
-    } else {
-      // just scan directory
-      let files = {};
-
-      const walk = async (dir) => {
-        const entries = await fs.promises.readdir(dir, { withFileTypes: true });
-        for( let entry of entries ) {
-          if( entry.isDirectory() ) {
-            await walk(path.join(dir, entry.name));
-          } else {
-            if( !files[type] ) files[type] = [];
-            if( !(type === 'cas' && relPath.endsWith('.json')) ) {
-              files[type].push(destPath); 
-            }
-          }
-        }
-      };
-      await walk(extractLocation);
-    }
-
-
-    }
-
-    if( opts.cb ) {
-      opts.cb({type: 'extract-complete'});
-    }
-
-    if( !files.cas ) files.cas = [];
-    files.files = [];
-    let lookup = new Map();
-
-    for( let file of files.cas ) { 
-      let metadata = JSON.parse(await fs.promises.readFile(file + '.json'));
-      metadata.files.forEach(f => {
-        f.hash = metadata.value;
-        delete f.fileId;
-        delete f.metadata.resourceType;
-        files.files.push(f);
-        lookup.set(path.join(f.directory, f.filename), f);
-      });
-    }
-
-    if( opts.cb ) {
-      opts.cb({ 
-        type: 'preflight-complete', 
-        stats: { 
-          totalFiles: files.files.length,
-          totalHashes: files.cas.length
-        } 
-      });
-    }
-
-    while( files.files.length > 0 ) {
-      let batch = files.files.splice(0, batchSize);
-
-      try {
-        let result = await cask.sync({requestor: 'admin'}, {
-          files: batch,
-          replace: opts.overwrite || false,
-        });
-        
-        stats.filesInserted += result.fileInserts.length;
-        stats.metadataUpdates += result.metadataUpdates.length;
-        stats.noChanges += result.noChanges.length;
-        stats.errors += result.errors?.length || 0;
-        result.errors.forEach(e => console.error(e));
-
-        let writtenHashes = new Set();
-        for( let file of result.doesNotExist ) {
-          let metadata = lookup.get(file);
-
-          let wContext = createContext({
-            requestor: 'admin',
-            filePath: file,
-            replace : opts.overwrite || false,
-            partitionKeys: metadata.partitionKeys,
-            metadata: metadata.metadata
-          }, db);
-
-          wContext.update({
-            replace : opts.overwrite || false,
-            partitionKeys: metadata.partitionKeys,
-            metadata: metadata.metadata
-          })
-
-          if( writtenHashes.has(metadata.hash) ) {
-            // This hash was already written in this batch, so we can skip writing the raw file again.
-            // Just update the context with the correct size for metadata processing and continue.
-            wContext.data.hash = metadata.hash;
-          } else {
-            // This hash has not been written yet, so we need to write the raw file from the extracted location.
-            wContext.data.readStream = fs.createReadStream(path.join(extractLocation, 'cas', this.cas._getHashFilePath(metadata.hash)));
-          }
-
-          await cask.write(wContext);
-          writtenHashes.add(metadata.hash);
-          stats.filesInserted += 1;
-        }
-
-        stats.filesProcessed += batch.length;
-        if( opts.cb ) {
-          opts.cb({ type: 'batch-sync', stats});
-        }
-      } catch (err) {
-        console.error(err);
-        stats.errors += 1;
-      }
-
-    }
-
-    if( fs.existsSync(extractLocation) ) {
-      await fs.promises.rm(extractLocation, { recursive: true });
-    }
-
-    return stats;
-  }
-
-  /**
-   * @method _extractArchiveToLocation
-   * @description Extract all tar.gz entries to extractLocation, preserving archive-relative paths.
-   *              extractLocation is the root directory for all extracted content.
-   *
-   * @param {stream.Readable} input - tar.gz readable stream
-   * @param {String} extractLocation - absolute destination root
-   * @returns {Promise<void>}
-   */
-  async _extractArchiveToLocation(input, extractLocation) {
-    const gunzip = zlib.createGunzip();
-    const extract = tarStream.extract();
-
-    let files = {}
-
-    extract.on('entry', async (header, stream, next) => {
-      try {
-        // keep archive layout rooted at extractLocation
-        const relPath = path.normalize(header.name).replace(/^(\.\.(\/|\\|$))+/, '');
-        const destPath = path.join(extractLocation, relPath);
-        const type = relPath.replace(/^\//,'').split('/')[0]; // 'cas', 'acl', or 'auto-partition'
-        if( !files[type] ) files[type] = [];
-        if( !(type === 'cas' && relPath.endsWith('.json')) ) {
-          files[type].push(destPath); 
-        }
-
-        // safety: prevent path traversal outside extractLocation
-        const root = path.resolve(extractLocation) + path.sep;
-        const resolvedDest = path.resolve(destPath);
-        if (!resolvedDest.startsWith(root) && resolvedDest !== path.resolve(extractLocation)) {
-          await this._drainStream(stream);
-          return next(new Error(`Invalid tar entry path: ${header.name}`));
-        }
-
-        if (header.type === 'directory') {
-          await fsp.mkdir(resolvedDest, { recursive: true });
-          await this._drainStream(stream);
-          return next();
-        }
-
-        if (header.type === 'file') {
-          await fsp.mkdir(path.dirname(resolvedDest), { recursive: true });
-          await pipeline(stream, fs.createWriteStream(resolvedDest));
-          return next();
-        }
-
-        // skip links and other special entries
-        await this._drainStream(stream);
-        next();
-      } catch (err) {
-        stream.resume();
-        next(err);
-      }
-    });
-
-    await pipeline(input, gunzip, extract);
-
-    return files;
-  }
-
-  /**
-   * @method _processImportEntry
-   * @description Dispatch a single tar entry to the appropriate import handler.
-   *
-   * @param {Object} header - tar entry header (name, size, …)
-   * @param {stream.Readable} stream - entry content stream; must be fully consumed
-   * @param {Object} opts - import options
-   * @param {Map} rawFileCache - shared map of hash → {size}
-   * @param {Object} aclData - accumulator for buffered ACL JSON
-   * @param {Object} autoPartitionData - accumulator for buffered auto-partition JSON
-   * @param {Object} summary - mutable import summary counters
-   * @param {import('./database/index.js').default} db - dedicated database connection
-   * @returns {Promise<void>}
-   */
-  async _processImportEntry(header, stream, opts, rawFileCache, aclData, autoPartitionData, summary, db) {
-    const name = header.name;
-
-    if (name.startsWith('cas/')) {
-      if (name.endsWith('.json')) {
-        const buf = await this._readStream(stream);
-        const meta = JSON.parse(buf.toString());
-        const fileInfo = rawFileCache.get(meta.value) || { size: 0 };
-        rawFileCache.delete(meta.value);
-
-        const result = await this._importCasMetadata(meta, opts, fileInfo, db);
-        console.log({result})
-        summary.fileCount += result.imported;
-        summary.skippedFiles += result.skipped;
-      } else {
-        const hash = path.basename(name);
-        const destPath = this.cas.diskPath(hash);
-        const { existed, size } = await this._writeCasRawFile(stream, destPath);
-        rawFileCache.set(hash, { size });
-        if (!existed) summary.hashCount++;
-      }
-    } else if (name.startsWith('acl/')) {
-      const buf = await this._readStream(stream);
-      const key = path.basename(name, '.json');
-      aclData[key] = JSON.parse(buf.toString());
-    } else if (name.startsWith('auto-partition/')) {
-      const buf = await this._readStream(stream);
-      const key = path.basename(name, '.json');
-      autoPartitionData[key] = JSON.parse(buf.toString());
-    } else {
-      await this._drainStream(stream);
-    }
-  }
-
   /**
    * @method _writeCasRawFile
    * @description Write a raw CAS file from the tar stream to its on-disk CAS location.
@@ -633,97 +338,6 @@ class Transfer {
     await pipeline(stream, fs.createWriteStream(destPath));
     const stat = await fsp.stat(destPath);
     return { existed: false, size: stat.size };
-  }
-
-  /**
-   * @method _importCasMetadata
-   * @description For a given hash's metadata object (parsed from the .json archive entry),
-   * insert or update all referenced file records in the database within a single transaction.
-   * RDF files are parsed and their triples inserted. The on-disk .json metadata file is
-   * regenerated from the DB state after all records are committed.
-   *
-   * @param {Object} meta - parsed .json metadata object from the archive
-   * @param {Object} opts - import options (overwrite, cb, …)
-   * @param {Object} fileInfo - {size} from the raw CAS file step
-   * @param {import('./database/index.js').default} db - dedicated database connection
-   * @returns {Promise<{imported: number, skipped: number}>}
-   */
-  async _importCasMetadata(meta, opts={}, fileInfo={}, db) {
-    let imported = 0;
-    let skipped = 0;
-
-    if (!meta.value || !meta.files || meta.files.length === 0) {
-      return { imported, skipped };
-    }
-
-    const hash = meta.value;
-    const casFilePath = this.cas.diskPath(hash);
-    const size = fileInfo.size || 0;
-    // Use the known primary digest; the hash column value IS the sha256 digest.
-    const digests = { [config.digests[0]]: hash };
-
-    try {
-      await db.query('BEGIN');
-
-      for (const fileRecord of meta.files) {
-        const filePath = path.join(fileRecord.directory, fileRecord.filename);
-        const fileExists = await db.fileExists(filePath);
-
-        if (fileExists) {
-          if (!opts.overwrite) {
-            throw new DuplicateFileError(filePath);
-          }
-          // Overwrite: update the existing record.
-          const existing = await db.getFile({ filePath });
-          const dirId = await this.directory.mkdir(fileRecord.directory, { dbClient: db });
-          await db.updateFile({
-            directoryId: dirId,
-            filePath,
-            hash,
-            metadata: fileRecord.metadata || {},
-            digests,
-            size,
-            user: 'import'
-          });
-          await db.clearFilePartitionKeys(existing.file_id);
-          for (const pk of (fileRecord.partitionKeys || [])) {
-            await db.addPartitionKeyToFile(existing.file_id, pk);
-          }
-          // Re-insert RDF if this is an RDF file.
-          const mimeType = (fileRecord.metadata || {}).mimeType;
-          if (RDF_MIME_TYPES.has(mimeType) || filePath.endsWith(JSONLD_EXT)) {
-            await this.rdf.delete(existing, { dbClient: db, ignoreAcl: true });
-            await this.rdf.insert(existing.file_id, { dbClient: db, filepath: casFilePath });
-          }
-        } else {
-          const dirId = await this.directory.mkdir(fileRecord.directory, { dbClient: db });
-          const fileId = await db.insertFile({
-            directoryId: dirId,
-            filePath,
-            hash,
-            metadata: fileRecord.metadata || {},
-            digests,
-            size,
-            user: 'import'
-          });
-          for (const pk of (fileRecord.partitionKeys || [])) {
-            await db.addPartitionKeyToFile(fileId, pk);
-          }
-          const mimeType = (fileRecord.metadata || {}).mimeType;
-          if (RDF_MIME_TYPES.has(mimeType) || filePath.endsWith(JSONLD_EXT)) {
-            await this.rdf.insert(fileId, { dbClient: db, filepath: casFilePath });
-          }
-          imported++;
-        }
-      }
-
-      await db.query('COMMIT');
-    } catch (err) {
-      await db.query('ROLLBACK');
-      throw err;
-    }
-
-    return { imported, skipped };
   }
 
   /**
