+++ src/schema/layer2-fs.sql
@@ -175,7 +175,57 @@ LEFT JOIN caskfs.directory_acl da ON d.directory_id = da.directory_id
 LEFT JOIN caskfs.root_directory_acl rda ON da.root_directory_acl_id = rda.root_directory_acl_id
 LEFT JOIN caskfs.directory rd ON rda.directory_id = rd.directory_id;
 
-CREATE MATERIALIZED VIEW IF NOT EXISTS caskfs.directory_user_permissions_lookup AS
+CREATE OR REPLACE FUNCTION caskfs.get_permission(
+  p_directory_id UUID,
+  p_user_id UUID DEFAULT NULL
+)
+RETURNS TABLE(can_read BOOLEAN, can_write BOOLEAN, is_admin BOOLEAN)
+LANGUAGE sql STABLE
+AS $$
+  -- Authenticated user check
+  SELECT
+    bool_or(rda.public OR p.permission IN ('read', 'write', 'admin')),
+    bool_or(p.permission IN ('write', 'admin')),
+    bool_or(p.permission = 'admin')
+  FROM caskfs.directory_acl da
+  JOIN caskfs.root_directory_acl rda USING (root_directory_acl_id)
+  LEFT JOIN caskfs.acl_permission p USING (root_directory_acl_id)
+  LEFT JOIN caskfs.acl_role_user ru ON p.role_id = ru.role_id AND ru.user_id = p_user_id
+  WHERE da.directory_id = p_directory_id
+    AND p_user_id IS NOT NULL
+
+  UNION ALL
+
+  -- Anonymous/public check
+  SELECT
+    bool_or(rda.public),
+    FALSE,
+    FALSE
+  FROM caskfs.directory_acl da
+  JOIN caskfs.root_directory_acl rda USING (root_directory_acl_id)
+  WHERE da.directory_id = p_directory_id
+    AND p_user_id IS NULL;
+$$;
+
+CREATE OR REPLACE FUNCTION caskfs.has_permission(
+  p_directory_id UUID,
+  p_permission caskfs.permission,
+  p_user_id UUID DEFAULT NULL
+)
+RETURNS BOOLEAN
+LANGUAGE sql STABLE
+AS $$
+  SELECT COALESCE(bool_or(
+    CASE p_permission
+      WHEN 'read' THEN can_read
+      WHEN 'write' THEN can_write
+      WHEN 'admin' THEN is_admin
+    END
+  ), FALSE)
+  FROM caskfs.get_permission(p_directory_id, p_user_id);
+$$;
+
+CREATE OR REPLACE VIEW caskfs.directory_user_permissions_lookup AS
 SELECT
     directory_id,
     user_id,
@@ -192,13 +242,13 @@ GROUP BY directory_id, user_id;
 ---
 
 -- Create indexes for the materialized view
-CREATE UNIQUE INDEX IF NOT EXISTS idx_directory_user_permissions_lookup_unique 
-    ON caskfs.directory_user_permissions_lookup(directory_id, user_id);
+-- CREATE UNIQUE INDEX IF NOT EXISTS idx_directory_user_permissions_lookup_unique 
+--     ON caskfs.directory_user_permissions_lookup(directory_id, user_id);
 -- This index is redundant because the unique index on (directory_id, user_id) already covers directory_id as the leftmost key.
 -- CREATE INDEX IF NOT EXISTS idx_directory_user_permissions_lookup_directory_id 
 --     ON caskfs.directory_user_permissions_lookup(directory_id);
-CREATE INDEX IF NOT EXISTS idx_directory_user_permissions_lookup_user_id 
-    ON caskfs.directory_user_permissions_lookup(user_id);
+-- CREATE INDEX IF NOT EXISTS idx_directory_user_permissions_lookup_user_id 
+--     ON caskfs.directory_user_permissions_lookup(user_id);
 
 CREATE OR REPLACE VIEW caskfs.directory_user_permissions_by_permission AS
 SELECT
